added verification gradle for @Size and @Email
