+++ src/main/java/org/launchcode/professionalprocrastinators/controllers/PlaylistController.java
@@ -1,0 +1,10 @@
+package org.launchcode.professionalprocrastinators.controllers;
+
+
+import org.springframework.stereotype.Controller;
+
+@Controller
+public class PlaylistController {
+
+
+}
